requirements.txt
streamlit
spacy
langdetect
deep-translator
reportlab
transformers 
torch
nltk
python-docx
PyMuPDF
